+++ ios/CDVBackgroundGeolocation/Config.m
@@ -10,7 +10,7 @@
 
 @implementation Config
 
-@synthesize stationaryRadius, distanceFilter, desiredAccuracy, isDebugging, activityType, stopOnTerminate, url, httpHeaders, saveBatteryOnBackground, maxLocations;
+@synthesize stationaryRadius, distanceFilter, desiredAccuracy, isDebugging, activityType, stopOnTerminate, url, syncUrl, syncThreshold, httpHeaders, saveBatteryOnBackground, maxLocations;
 
 -(id) init {
     self = [super init];
@@ -27,6 +27,7 @@
     stopOnTerminate = NO;
     saveBatteryOnBackground = YES;
     maxLocations = 10000;
+    syncThreshold = 100;
     
     return self;
 }
@@ -36,13 +37,13 @@
     Config *instance = [[Config alloc] init];
 
     if (config[@"stationaryRadius"]) {
-        instance.stationaryRadius = [config[@"stationaryRadius"] intValue];
+        instance.stationaryRadius = [config[@"stationaryRadius"] integerValue];
     }
     if (config[@"distanceFilter"]) {
-        instance.distanceFilter = [config[@"distanceFilter"] intValue];
+        instance.distanceFilter = [config[@"distanceFilter"] integerValue];
     }
     if (config[@"desiredAccuracy"]) {
-        instance.desiredAccuracy = [config[@"desiredAccuracy"] intValue];
+        instance.desiredAccuracy = [config[@"desiredAccuracy"] integerValue];
     }
     if (config[@"debug"]) {
         instance.isDebugging = [config[@"debug"] boolValue];
@@ -56,6 +57,14 @@
     if (config[@"url"]) {
         instance.url = config[@"url"];
     }
+    if (config[@"syncUrl"]) {
+        instance.syncUrl = config[@"syncUrl"];
+    } else if (config[@"url"]) {
+        instance.syncUrl = config[@"url"];
+    }
+    if (config[@"syncThreshold"]) {
+        instance.syncThreshold = [config[@"syncThreshold"] integerValue];
+    }
     if (config[@"httpHeaders"]) {
         instance.httpHeaders = config[@"httpHeaders"];
     }
@@ -63,7 +72,7 @@
         instance.saveBatteryOnBackground = [config[@"saveBatteryOnBackground"] boolValue];
     }
     if (config[@"maxLocations"]) {
-        instance.maxLocations = [config[@"maxLocations"] intValue];
+        instance.maxLocations = [config[@"maxLocations"] integerValue];
     }
 
     return instance;