+++ ios/CDVBackgroundGeolocation/Location.h
@@ -37,6 +37,7 @@
 - (BOOL) isBetterLocation:(Location*)location;
 - (BOOL) isBeyond:(Location*)location radius:(NSInteger)radius;
 - (BOOL) isValid;
+- (BOOL) postAsJSON:(NSString*)url withHttpHeaders:(NSMutableDictionary*)httpHeaders error:(NSError * __autoreleasing *)outError;
 - (id) copyWithZone: (NSZone *)zone;
 
 @end