+++ ios/CDVBackgroundGeolocationTests/SQLiteLocationDAOTests.m
@@ -24,8 +24,6 @@
 - (void)tearDown {
     // Put teardown code here. This method is called after the invocation of each test method in the class.
     [super tearDown];
-    SQLiteLocationDAO *locationDAO = [SQLiteLocationDAO sharedInstance];
-    [locationDAO clearDatabase];
 }
 
 - (void)testPersistLocation {