+++ ios/CDVBackgroundGeolocationTests/SQLiteLocationDAOThreadTests.m
@@ -34,8 +34,6 @@
 - (void)tearDown {
     // Put teardown code here. This method is called after the invocation of each test method in the class.
     [super tearDown];
-    SQLiteLocationDAO *locationDAO = [SQLiteLocationDAO sharedInstance];
-    [locationDAO clearDatabase];
 }
 
 - (void)testGetAllLocationsMultiThread {
@@ -83,13 +81,6 @@
         XCTAssertTrue([result.service_provider isEqualToNumber:[NSNumber numberWithInt:-1]], "service_provider is %@ expecting %@", result.service_provider, [NSNumber numberWithInt:-1]);
         XCTAssertEqual([result debug], 1, @"debug is expected to be true");
     }
-}
-
-- (void)testPerformanceExample {
-    // This is an example of a performance test case.
-    [self measureBlock:^{
-        // Put the code you want to measure the time of here.
-    }];
 }
 
 @end